+++ src/pages/Account/User.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { Form, Icon, Input, Button, Col , Menu, Row, Card,Alert,notification
+import { Form, Icon, Input, Button, Col, Menu, Row, Card, Alert, notification
   // Checkbox
   } from 'antd';
 import IconFont from "../../components/IconFont";
@@ -15,12 +15,12 @@ class User extends React.Component {
 
     constructor(props) {
         super(props);
-        this.state = {
-            menuindex: 0,
-            oldPassword: '',
-            newPassword: '',
-            confirmNewPassword: ''
-        }
+        // this.state = {
+        //     menuindex: 0,
+        //     oldPassword: '',
+        //     newPassword: '',
+        //     confirmNewPassword: ''
+        // }
     }
 
     menuClicked(e) {
@@ -46,81 +46,81 @@ class User extends React.Component {
       }
     }
 
-    handlePasswordChange(event, pwtype) {
-      const obj = {}
-      obj[pwtype] = event.target.value
-      this.setState(obj);
-    }
+    // handlePasswordChange(event, pwtype) {
+    //   const obj = {}
+    //   obj[pwtype] = event.target.value
+    //   this.setState(obj);
+    // }
 
-    async changePassword(e) {
-      const { oldPassword, newPassword, confirmNewPassword } = this.state
-      if (newPassword !== confirmNewPassword) {
-        notification.error({
-          message: '两次新密码不匹配',
-          description: '请重新输入新密码.'
-        })
-      } else {
-        try {
-          await changePassword({ old_password:oldPassword, new_password:newPassword })
-          notification.success({
-            message: '密码修改成功'
-          })
-        } catch (error) {
-          notification.error({
-            message: error.message
-          })
-        }
-      }
-    }
+    // async changePassword(e) {
+    //   const { oldPassword, newPassword, confirmNewPassword } = this.state
+    //   if (newPassword !== confirmNewPassword) {
+    //     notification.error({
+    //       message: '两次新密码不匹配',
+    //       description: '请重新输入新密码.'
+    //     })
+    //   } else {
+    //     try {
+    //       await changePassword({ old_password:oldPassword, new_password:newPassword })
+    //       notification.success({
+    //         message: '密码修改成功'
+    //       })
+    //     } catch (error) {
+    //       notification.error({
+    //         message: error.message
+    //       })
+    //     }
+    //   }
+    // }
 
-    renderInfo() {
-        if(this.state.menuindex === 0) {
-        return <Form style={ style.container }>
-                <h2>个人信息</h2>
-                <br />
-                <FormItem>
-                    <Input prefix={<Icon type="user" style={{ color: 'rgba(0,0,0,.25)' }} />} placeholder="用户名" />
-                </FormItem>
-                <FormItem>
-                    <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} placeholder="邮箱" />
-                </FormItem>
-                <FormItem>
-                  <Button type="primary" htmlType="submit" className="login-form-button">
-                    确认修改
-                  </Button>
-                </FormItem>
-              </Form>
-            }
-    }
+    // renderInfo() {
+    //     if(this.state.menuindex === 0) {
+    //     return <Form style={ style.container }>
+    //             <h2>个人信息</h2>
+    //             <br />
+    //             <FormItem>
+    //                 <Input prefix={<Icon type="user" style={{ color: 'rgba(0,0,0,.25)' }} />} placeholder="用户名" />
+    //             </FormItem>
+    //             <FormItem>
+    //                 <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} placeholder="邮箱" />
+    //             </FormItem>
+    //             <FormItem>
+    //               <Button type="primary" htmlType="submit" className="login-form-button">
+    //                 确认修改
+    //               </Button>
+    //             </FormItem>
+    //           </Form>
+    //         }
+    // }
 
-    renderPassword() {
-        if(this.state.menuindex === 1) {
-        return <Form style={ style.container }>
-                <h2>修改密码</h2>
-                <br />
-                <FormItem>
-                    <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} type="password" placeholder="旧密码" 
-                    value={this.state.oldPassword}
-                    onChange={(e) => this.handlePasswordChange(e, 'oldPassword')}/>
-                </FormItem>
-                <FormItem>
-                    <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} type="password" placeholder="新密码" 
-                    value={this.state.newPassword}
-                    onChange={(e) => this.handlePasswordChange(e, 'newPassword')}/>
-                </FormItem>
-                <FormItem>
-                    <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} type="password" placeholder="确认新密码" 
-                    value={this.state.confirmNewPassword}
-                    onChange={(e) => this.handlePasswordChange(e, 'confirmNewPassword')}/>
-                </FormItem>
-                <FormItem>
-                  <Button type="primary" htmlType="submit" className="login-form-button" onClick={this.changePassword}>
-                    确认修改
-                  </Button>
-                </FormItem>
-              </Form>
-            }
-    }
+    // renderPassword() {
+    //     if(this.state.menuindex === 1) {
+    //     return <Form style={ style.container }>
+    //             <h2>修改密码</h2>
+    //             <br />
+    //             <FormItem>
+    //                 <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} type="password" placeholder="旧密码" 
+    //                 value={this.state.oldPassword}
+    //                 onChange={(e) => this.handlePasswordChange(e, 'oldPassword')}/>
+    //             </FormItem>
+    //             <FormItem>
+    //                 <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} type="password" placeholder="新密码" 
+    //                 value={this.state.newPassword}
+    //                 onChange={(e) => this.handlePasswordChange(e, 'newPassword')}/>
+    //             </FormItem>
+    //             <FormItem>
+    //                 <Input prefix={<Icon type="lock" style={{ color: 'rgba(0,0,0,.25)' }} />} type="password" placeholder="确认新密码" 
+    //                 value={this.state.confirmNewPassword}
+    //                 onChange={(e) => this.handlePasswordChange(e, 'confirmNewPassword')}/>
+    //             </FormItem>
+    //             <FormItem>
+    //               <Button type="primary" htmlType="submit" className="login-form-button" onClick={this.changePassword}>
+    //                 确认修改
+    //               </Button>
+    //             </FormItem>
+    //           </Form>
+    //         }
+    // }
 
     renderWallet() {
         if(this.state.menuindex === 2) {
